+++ include/text_util.h
@@ -3,7 +3,7 @@
 
 void sub_80922B4(u8 *buffer, const u8 *string, s32 size);
 u32 sub_80922E4(u32 r0);
-void CopyStringtoBuffer(u8 *buffer, u8 *string);
+void CopyStringtoBuffer(u8 *buffer, const u8 *string);
 void BoundedCopyStringtoBuffer(u8 *buffer, u8 *string, s32 size);
 const char *GetUnformattedTypeString(u8 type);
 const char * GetFormattedTypeString(u8 type);
