+++ src/dungeon_range.c
@@ -259,3 +259,5 @@ EntityInfo* GetLeaderInfo(void)
 {
     return GetEntInfo(GetLeader());
 }
+
+// all of code_8083654.s belongs to this file
